sync with original master

+++ tensorflow/python/util/tf_stack.cc
@@ -19,8 +19,8 @@ limitations under the License.
 #include <algorithm>
 #include <vector>
 
-#include "include/pybind11/pybind11.h"
-#include "include/pybind11/stl_bind.h"
+#include "pybind11/pybind11.h"
+#include "pybind11/stl_bind.h"
 
 struct FrameSummary;  // Forward declaration.
 
@@ -83,7 +83,8 @@ std::vector<FrameSummary> ExtractStack(ssize_t limit, const py::list& mappers,
   std::vector<FrameSummary> ret;
   // 16 is somewhat arbitrary, but TensorFlow stack traces tend to be deep.
   ret.reserve(limit < 0 ? 16 : static_cast<size_t>(limit));
-  for (; f != nullptr && (limit < 0 || ret.size() < limit); f = f->f_back) {
+  for (; f != nullptr && (limit < 0 || ret.size() < static_cast<size_t>(limit));
+       f = f->f_back) {
     const PyCodeObject* co = f->f_code;
     int lineno = PyFrame_GetLineNumber(const_cast<PyFrameObject*>(f));
     auto filename = py::reinterpret_borrow<py::str>(co->co_filename);
